requirements.txt
langchain==1.2.0
flask==3.1.1
sentence-transformers==4.1.0 
pypdf==5.6.1              
langchain-classic==1.0.0            
langchain-community==0.4.1  
python-dotenv==1.1.0                     
langchain-google-genai==4.1.2            
langchain-huggingface==1.2.0            
langchain-pinecone==0.2.8  
langchain-text-splitters==1.1.0 
 -e .